Cleanup

+++ src/deemon/method-hints/seq_all.h
@@ -114,7 +114,7 @@ int __seq_all__.seq_all([[nonnull]] DeeObject *__restrict self)
 	Dee_ssize_t foreach_status;
 	foreach_status = CALL_DEPENDENCY(seq_operator_foreach, self, &seq_all_foreach_cb, NULL);
 	ASSERT(foreach_status == 0 || foreach_status == -1 || foreach_status == -2);
-	return Dee_HAS_FROM_eM1_nM2_y0(foreach_status);
+	return Dee_HAS_FROM_eM1_nM2_y0((int)foreach_status);
 }} {
 	DREF DeeObject *result = LOCAL_CALLATTR(self, 0, NULL);
 	if unlikely(!result)
@@ -151,7 +151,7 @@ int __seq_all__.seq_all_with_key([[nonnull]] DeeObject *self,
 	Dee_ssize_t foreach_status;
 	foreach_status = CALL_DEPENDENCY(seq_operator_foreach, self, &seq_all_with_key_foreach_cb, key);
 	ASSERT(foreach_status == 0 || foreach_status == -1 || foreach_status == -2);
-	return Dee_HAS_FROM_eM1_nM2_y0(foreach_status);
+	return Dee_HAS_FROM_eM1_nM2_y0((int)foreach_status);
 }} {
 	DREF DeeObject *result;
 	DeeObject *args[3];
@@ -190,7 +190,7 @@ int __seq_all__.seq_all_with_range([[nonnull]] DeeObject *__restrict self,
 	Dee_ssize_t foreach_status;
 	foreach_status = CALL_DEPENDENCY(seq_enumerate_index, self, &seq_all_enumerate_cb, NULL, start, end);
 	ASSERT(foreach_status == 0 || foreach_status == -1 || foreach_status == -2);
-	return Dee_HAS_FROM_eM1_nM2_y0(foreach_status);
+	return Dee_HAS_FROM_eM1_nM2_y0((int)foreach_status);
 }} {
 	DREF DeeObject *result = LOCAL_CALLATTRF(self, PCKuSIZ PCKuSIZ, start, end);
 	if unlikely(!result)
@@ -225,7 +225,7 @@ int __seq_all__.seq_all_with_range_and_key([[nonnull]] DeeObject *self,
 	Dee_ssize_t foreach_status;
 	foreach_status = CALL_DEPENDENCY(seq_enumerate_index, self, &seq_all_with_key_enumerate_cb, key, start, end);
 	ASSERT(foreach_status == 0 || foreach_status == -1 || foreach_status == -2);
-	return Dee_HAS_FROM_eM1_nM2_y0(foreach_status);
+	return Dee_HAS_FROM_eM1_nM2_y0((int)foreach_status);
 }} {
 	DREF DeeObject *result = LOCAL_CALLATTRF(self, PCKuSIZ PCKuSIZ "o", start, end, key);
 	if unlikely(!result)
