+++ src/deemon/method-hints/seq_any.h
@@ -113,7 +113,7 @@ int __seq_any__.seq_any([[nonnull]] DeeObject *__restrict self)
 	Dee_ssize_t foreach_status;
 	foreach_status = CALL_DEPENDENCY(seq_operator_foreach, self, &seq_any_foreach_cb, NULL);
 	ASSERT(foreach_status == 0 || foreach_status == -1 || foreach_status == -2);
-	return Dee_HAS_FROM_eM1_n0_yM2(foreach_status);
+	return Dee_HAS_FROM_eM1_n0_yM2((int)foreach_status);
 }} {
 	DREF DeeObject *result = LOCAL_CALLATTR(self, 0, NULL);
 	if unlikely(!result)
@@ -149,7 +149,7 @@ int __seq_any__.seq_any_with_key([[nonnull]] DeeObject *self,
 	Dee_ssize_t foreach_status;
 	foreach_status = CALL_DEPENDENCY(seq_operator_foreach, self, &seq_any_with_key_foreach_cb, key);
 	ASSERT(foreach_status == 0 || foreach_status == -1 || foreach_status == -2);
-	return Dee_HAS_FROM_eM1_n0_yM2(foreach_status);
+	return Dee_HAS_FROM_eM1_n0_yM2((int)foreach_status);
 }} {
 	DREF DeeObject *result;
 	DeeObject *args[3];
@@ -200,7 +200,7 @@ err:
 	Dee_ssize_t foreach_status;
 	foreach_status = CALL_DEPENDENCY(seq_enumerate_index, self, &seq_any_enumerate_cb, NULL, start, end);
 	ASSERT(foreach_status == 0 || foreach_status == -1 || foreach_status == -2);
-	return Dee_HAS_FROM_eM1_n0_yM2(foreach_status);
+	return Dee_HAS_FROM_eM1_n0_yM2((int)foreach_status);
 }} {
 	DREF DeeObject *result = LOCAL_CALLATTRF(self, PCKuSIZ PCKuSIZ, start, end);
 	if unlikely(!result)
@@ -233,7 +233,7 @@ int __seq_any__.seq_any_with_range_and_key([[nonnull]] DeeObject *self, size_t s
 	Dee_ssize_t foreach_status;
 	foreach_status = CALL_DEPENDENCY(seq_enumerate_index, self, &seq_any_with_key_enumerate_cb, key, start, end);
 	ASSERT(foreach_status == 0 || foreach_status == -1 || foreach_status == -2);
-	return Dee_HAS_FROM_eM1_n0_yM2(foreach_status);
+	return Dee_HAS_FROM_eM1_n0_yM2((int)foreach_status);
 }} {
 	DREF DeeObject *result = LOCAL_CALLATTRF(self, PCKuSIZ PCKuSIZ "o", start, end, key);
 	if unlikely(!result)
